task/chrome-extension-optimization Optimized the code structure and made it make readable for the different events to handle
