+++ digizuite-chrome-extension/src/utility/options.js
@@ -1,0 +1,3 @@
+export function openSettings() {
+    chrome.runtime.openOptionsPage();
+}
